Configuration file prefix cleanup, using global for conf and lib directory, loading ssl paths and seperate directory
