+++ src/deepstream.io.js
@@ -7,10 +7,10 @@ var ConnectionEndpoint = require( './message/connection-endpoint' ),
 	messageParser = require( './message/message-parser' ),
 	readMessage = require( './utils/read-message' ),
 	util = require( 'util' ),
-	configInitialiser = require( './utils/config-initialiser' ),
 	utils = require( './utils/utils' ),
 	defaultOptions = require( './default-options' ),
-	jsYamlLoader = require( './utils/js-yaml-loader' ),
+	configInitialiser = require( './config/config-initialiser' ),
+	jsYamlLoader = require( './config/js-yaml-loader' ),
 	RpcHandler = require( './rpc/rpc-handler' ),
 	RecordHandler = require( './record/record-handler' ),
 	WebRtcHandler = require( './webrtc/webrtc-handler' ),
