+++ src/utils/utils.js
@@ -173,53 +173,3 @@ exports.merge = function() {
 
 	return result;
 };
-
-/**
- * If a prefix is not set the filePath will be returned
- *
- * Otherwise it will either replace return a new path prepended with the prefix.
- * If the prefix is not an absolute path it will also prepend the CWD.
- *
- * file        || relative (starts with .) | absolute | else (npm module path)
- * -----------------------------------------------------------------------------
- * *prefix     || *CWD + prefix + file     | file     | *CWD + prefix + file
- * *no prefix  ||  CWD + file              | file     | file (resolved by nodes require)
- *
- * *CWD = ignore CWD if prefix is absolute
- *
- * @param {String} filePath
- * @param {String} prefix
- *
- * @private
- * @returns {String} file path with the prefix
- */
-exports.lookupRequirePath = function( filePath, prefix ) {
-	if ( path.parse( filePath ).root !== '' ) {
-		// filePath is absolute
-		return filePath;
-	} else if ( filePath[0] !== '.' ) {
-		// filePath is not relative (and not absolute)
-		if ( prefix == null ) {
-			return filePath;
-		} else {
-			return resolvePrefixAndFile( filePath, prefix );
-		}
-	} else {
-		// filePath is relative, starts with .
-		if ( prefix == null ) {
-			return path.resolve( process.cwd(), filePath );
-		} else {
-			return resolvePrefixAndFile( filePath, prefix );
-		}
-	}
-};
-
-function resolvePrefixAndFile( nonAbsoluteFilePath, prefix ) {
-	if ( path.parse( prefix ).root === '' ) {
-		// prefix is not absolute
-		return path.resolve( process.cwd(), prefix, nonAbsoluteFilePath );
-	} else {
-		// prefix is absolute
-		return path.resolve( prefix, nonAbsoluteFilePath );
-	}
-}
