+++ test/utils/utilsSpec.js
@@ -293,40 +293,3 @@ describe( 'merges recoursively', function(){
 
 	});
 });
-
-describe( 'lookupRequirePath', function() {
-	it( 'check cases with no or a relative prefix', function() {
-		// node style path (no dot at the start and not absolute path)
-		expect( utils.lookupRequirePath( 'foo-bar' ) ).toEqual( 'foo-bar' );
-		expect( utils.lookupRequirePath( 'dir/foo-bar' ) ).toEqual( 'dir/foo-bar' );
-		expect( utils.lookupRequirePath( 'foo-bar', 'pre' ) ).toEqual( path.resolve( 'pre', 'foo-bar' ) );
-		expect( utils.lookupRequirePath( 'dir/foo-bar', 'pre' ) ).toEqual( path.resolve( 'pre', 'dir', 'foo-bar' ) );
-
-		// use an absolute path for the filename
-		expect( utils.lookupRequirePath( '/usr/foo-bar' ) ).toEqual( '/usr/foo-bar' );
-		expect( utils.lookupRequirePath( '/usr/dir/foo-bar' ) ).toEqual( '/usr/dir/foo-bar' );
-		expect( utils.lookupRequirePath( '/usr/foo-bar', 'pre' ) ).toEqual( '/usr/foo-bar' );
-		expect( utils.lookupRequirePath( '/usr/dir/foo-bar', 'pre' ) ).toEqual( '/usr/dir/foo-bar' );
-
-		// use a relative path for the filename
-		expect( utils.lookupRequirePath( './foo-bar' ) ).toEqual( path.resolve( 'foo-bar' ) );
-		expect( utils.lookupRequirePath( './dir/foo-bar' ) ).toEqual( path.resolve( 'dir', 'foo-bar' ) );
-		expect( utils.lookupRequirePath( './foo-bar', 'pre' ) ).toEqual( path.resolve( 'pre', 'foo-bar' ) );
-		expect( utils.lookupRequirePath( './dir/foo-bar', 'pre' ) ).toEqual( path.resolve( 'pre', 'dir', 'foo-bar' ) );
-	} );
-
-	it( 'check cases with an absolute prefix', function() {
-		// node style path (no dot at the start and not absolute path)
-		expect( utils.lookupRequirePath( 'foo-bar', '/pre' ) ).toEqual( path.resolve( '/pre', 'foo-bar' ) );
-		expect( utils.lookupRequirePath( 'dir/foo-bar', '/pre' ) ).toEqual( path.resolve( '/pre', 'dir', 'foo-bar' ) );
-
-		// use an absolute path for the filename
-		expect( utils.lookupRequirePath( '/usr/foo-bar', '/pre' ) ).toEqual( '/usr/foo-bar' );
-		expect( utils.lookupRequirePath( '/usr/dir/foo-bar', '/pre' ) ).toEqual( '/usr/dir/foo-bar' );
-
-		// use a relative path for the filename
-		expect( utils.lookupRequirePath( './foo-bar', '/pre' ) ).toEqual( path.resolve( '/pre', 'foo-bar' ) );
-		expect( utils.lookupRequirePath( './dir/foo-bar', '/pre' ) ).toEqual( path.resolve( '/pre', 'dir', 'foo-bar' ) );
-
-	} );
-} );
